Fixed a bunch of Key related issues
